pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm_escalation_evaluator"
version = "0.1.0"
description = "Nurse-patient escalation grading tool using OpenAI Structured Outputs."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai>=1.0.0",
  "python-dotenv>=1.0.0"
]

[project.scripts]
escalation-grader = "escalation_grader.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]